feat(core/shmem): move shared memory logic to dedicated source file
This patch relocates the initialization and shared memory access steps
from the hardcoded settings in the existing IPC file to a generic
shared memory file. Previously, the code assumed the existence of
only one object working with the shared memory (ipc). With this update,
the system is now prepared for future versions where multiple objects
can be added and working with the shared memory.
