+++ src/core/objects.mk
@@ -13,3 +13,4 @@ core-objs-y+=console.o
 core-objs-y+=ipc.o
 core-objs-y+=objpool.o
 core-objs-y+=hypercall.o
+core-objs-y+=shmem.o
